add authority, version, and attachments files, remove corresponding tests
