+++ test/v1_0_3/non_templating.js
@@ -17,233 +17,6 @@
     if(global.OAUTH)
         request = helper.OAuthRequest(request);
 
-    describe('An LRS populates the "authority" property if it is not provided in the Statement, based on header information with the Agent corresponding to the user (contained within the header) (Implicit, 4.1.9.b, 4.1.9.c) ', function () {
-
-        it('should populate authority ', function (done) {
-
-            this.timeout(0);
-            var templates = [
-                {statement: '{{statements.default}}'}
-            ];
-            var data = createFromTemplate(templates);
-            data = data.statement;
-            data.id = helper.generateUUID();
-            var query = '?statementId=' + data.id;
-            var stmtTime = Date.now();
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders({}))
-                .json(data)
-                .expect(200)
-                .end()
-                .get(helper.getEndpointStatements() + query)
-                .headers(helper.addAllHeaders({}))
-                .wait(genDelay(stmtTime, query, data.id))
-                .expect(200).end(function (err, res) {
-                    if (err) {
-                        done(err);
-                    } else {
-                        var statement = parse(res.body, done);
-                        expect(statement).to.have.property('authority');
-                        done();
-                    }
-                });
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a Request which uses Attachments and does not have a "Content-Type" header with value "application/json" or "multipart/mixed" (Format, 4.1.11.a, 4.1.11.b)', function () {
-        var data;
-        var attachment;
-
-        before('create attachment templates', function () {
-            var templates = [
-                {statement: '{{statements.attachment}}'},
-                {
-                    attachments: [
-                        {
-                            "usageType": "http://example.com/attachment-usage/test",
-                            "display": {"en-US": "A test attachment"},
-                            "description": {"en-US": "A test attachment (description)"},
-                            "contentType": "text/plain; charset=ascii",
-                            "length": 27,
-                            "sha2": "495395e777cd98da653df9615d09c0fd6bb2f8d4788394cd53c56a3bfdcd848a",
-                            "fileUrl": "http://over.there.com/file.txt"
-                        }
-                    ]
-                }
-            ];
-            data = createFromTemplate(templates);
-            data = data.statement;
-
-            attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_image_multipart_attachment_valid.part', {encoding: 'binary'});
-        });
-
-        it('should succeed when attachment uses "fileUrl" and request content-type is "application/json"', function (done) {
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders({}))
-                .json(data).expect(200, done);
-        });
-
-        it('should fail when attachment uses "fileUrl" and request content-type is "multipart/form-data"', function (done) {
-            var header = {'Content-Type': 'multipart/form-data; boundary=-------314159265358979323846'};
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(JSON.stringify(data)).expect(400, done);
-        });
-
-        it('should succeed when attachment is raw data and request content-type is "multipart/mixed"', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(200, done);
-        });
-
-        it('should fail when attachment is raw data and request content-type is "multipart/form-data"', function (done) {
-            var header = {'Content-Type': 'multipart/form-data; boundary=-------314159265358979323846'};
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content-Type" header with value "application/json", and has a discrepancy in the number of Attachments vs. the number of fileURL members (4.1.11.a)', function () {
-        it('should fail when passing statement attachments and missing attachment"s binary', function (done) {
-            var templates = [
-                {statement: '{{statements.attachment}}'},
-                {
-                    attachments: [
-                        {
-                            "usageType": "http://example.com/attachment-usage/test",
-                            "display": {"en-US": "A test attachment"},
-                            "description": {"en-US": "A test attachment (description)"},
-                            "contentType": "text/plain; charset=ascii",
-                            "length": 27,
-                            "sha2": "495395e777cd98da653df9615d09c0fd6bb2f8d4788394cd53c56a3bfdcd848a",
-                            "fileUrl": "http://over.there.com/file.txt"
-                        },
-                        {
-                            "usageType": "http://example.com/attachment-usage/test",
-                            "display": {"en-US": "A test attachment"},
-                            "description": {"en-US": "A test attachment (description)"},
-                            "contentType": "text/plain; charset=ascii",
-                            "length": 27,
-                            "sha2": "495395e777cd98da653df9615d09c0fd6bb2f8d4788394cd53c56a3bfdcd848a"
-                        }
-                    ]
-                }
-            ];
-            var data = createFromTemplate(templates);
-            data = data.statement;
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders({}))
-                .json(data).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and does not have a body header named "Content-Type" with value "multipart/mixed" (RFC 1341)', function () {
-        it('should fail when attachment is raw data and first part content type is not "application/json"', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_first_part_content_type.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and does not have a body header named "boundary" (4.1.11.b, RFC 1341)', function () {
-        it('should fail if boundary not provided in body', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_first_part_no_boundary.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('A Boundary is defined as the value of the body header named "boundary" (Definition, 4.1.11.b, RFC 1341)', function () {
-        it('should fail if boundary not provided in header', function (done) {
-            var header = {'Content-Type': 'multipart/mixed;'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_valid.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and does not have a Boundary before each "Content-Type" header (4.1.11.b, RFC 1341)', function () {
-        it('should fail if boundary not provided in body', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_first_part_no_boundary.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and does not the first document part with a "Content-Type" header with a value of "application/json" (RFC 1341, 4.1.11.b.a)', function () {
-        it('should fail when attachment is raw data and first part content type is not "application/json"', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_first_part_content_type.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and does not have all of the Statements in the first document part (RFC 1341, 4.1.11.b.a)', function () {
-        it('should fail when statements separated into multiple parts', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_statement_parts.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
-    describe('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and for any part except the first does not have a Header named "X-Experience-API-Hash" with a value of one of those found in a "sha2" property of a Statement in the first part of this document (4.1.11.b.c, 4.1.11.b.d)', function () {
-        it('should fail when attachments missing header "X-Experience-API-Hash"', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_no_x_experience_api_hash.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-
-        it('should fail when attachments header "X-Experience-API-Hash" does not match "sha2"', function (done) {
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_no_match_sha2.part', {encoding: 'binary'});
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(400, done);
-        });
-    });
-
     describe('An LRS rejects with error code 400 Bad Request, a Request which does not use a "X-Experience-API-Version" header name to any API except the About API (Format, 6.2.a, 6.2.f, 7.7.f)', function () {
         it('should pass when GET without header "X-Experience-API-Version"', function (done) {
             request(helper.getEndpointAndAuth())
@@ -3290,153 +3063,6 @@
               });
         });
 
-        it('An LRS rejects with error code 400 Bad Request, a GET Request which uses Attachments, has a "Content-Type" header with value "application/json", and has the "attachments" filter attribute set to "true" (4.1.11.a)', function (done) {
-
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_image_multipart_attachment_valid.part', {encoding: 'binary'});
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-
-            var data = {
-                attachments: true,
-                limit: 1
-            };
-            var query = helper.getUrlEncoding(data);
-            var stmtTime = Date.now();
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(200)
-                .end()
-                .get(helper.getEndpointStatements() + '?' + query)
-                .wait(genDelay(stmtTime, '?' + query, null))
-                .headers(helper.addAllHeaders(header))
-                .expect(200)
-                .end(function(err,res){
-                    if (err){
-                        done(err);
-                    }
-                    else{
-                        expect(res.headers['content-type']).to.have.string('multipart/mixed');
-                        done();
-                    }
-                });
-              });
-
-        it('An LRS\'s Statement API will reject a GET request having the "attachment" parameter set to "false" and the Content-Type field in the header set to anything but "application/json" (7.2.3.d, 7.2.3.e)', function (done) {
-
-            var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_image_multipart_attachment_valid.part', {encoding: 'binary'});
-            var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-
-            var data = {
-                attachments: false,
-                limit: 1
-            };
-            var query = helper.getUrlEncoding(data);
-            var stmtTime = Date.now();
-
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders(header))
-                .body(attachment).expect(200)
-                .end()
-                .get(helper.getEndpointStatements() + '?' + query)
-                .wait(genDelay(stmtTime, '?' + query, null))
-                .headers(helper.addAllHeaders(header))
-                .expect(200)
-                .end(function(err,res){
-                    if (err){
-                        done(err);
-                    }
-                    else{
-                        expect(res.headers['content-type']).to.equal('application/json');
-                        done();
-                    }
-                });
-        });
-
-        it('An LRS rejects with error code 400 Bad Request, a PUT or POST Request which uses Attachments, has a "Content Type" header with value "multipart/mixed", and for any part except the first does not have a Header named "Content-Transfer-Encoding" with a value of "binary" (4.1.11.b.c, 4.1.11.b.e)', function (done) {
-
-          var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_text_multipart_attachment_invalid_content_transfer_encoding.part', {encoding: 'binary'});
-          var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-
-          request(helper.getEndpointAndAuth())
-              .post(helper.getEndpointStatements())
-              .headers(helper.addAllHeaders(header))
-              .body(attachment).expect(400,done);
-        });
-
-        it ('An LRS\'s Statement API will reject a GET request having the "attachment" parameter set to "true" if it does not follow the rest of the attachment rules (7.2.3.d)', function (done){
-          //incomplete- should compare raw data between request and response
-          var attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_image_multipart_attachment_valid.part', {encoding: 'binary'});
-          var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-          var stmtTime = Date.now();
-
-          var data = {
-              attachments: true,
-              limit: 1
-          };
-          var query = helper.getUrlEncoding(data);
-
-          request(helper.getEndpointAndAuth())
-              .post(helper.getEndpointStatements())
-              .headers(helper.addAllHeaders(header))
-              .body(attachment).expect(200)
-              .end()
-              .get(helper.getEndpointStatements()+ '?' + query)
-              .wait(genDelay(stmtTime, '?' + query, null))
-              .headers(helper.addAllHeaders(header))
-              .expect(200)
-              .end(function(err,res){
-                  if (err){
-                      done(err);
-                  }
-                  else{
-                      try {
-                          var request = JSON.parse(res.body);
-                          done("Contains raw data");
-                      } catch (e) {
-                          expect(res.headers['content-type']).to.have.string('multipart/mixed');
-                          done();
-                      }
-
-                  }
-              });
-        });
-
-        it ('An LRS\'s Statement API will reject a GET request having the "attachment" parameter set to "false" if it includes attachment raw data (7.2.3.d)', function (done){
-
-              var data = {
-                  attachments: false,
-                  limit: 1
-              };
-              var query = helper.getUrlEncoding(data);
-
-              attachment = fs.readFileSync('test/v1_0_3/templates/attachments/basic_image_multipart_attachment_valid.part', {encoding: 'binary'});
-              var header = {'Content-Type': 'multipart/mixed; boundary=-------314159265358979323846'};
-              var stmtTime = Date.now();
-
-              request(helper.getEndpointAndAuth())
-                  .post(helper.getEndpointStatements())
-                  .headers(helper.addAllHeaders(header))
-                  .body(attachment).expect(200)
-                  .end()
-                  .get(helper.getEndpointStatements()+ '?' + query)
-                  .wait(genDelay(stmtTime, '?' + query, null))
-                  .headers(helper.addAllHeaders(header))
-                  .expect(200)
-                  .end(function(err,res){
-                      if (err){
-                          done(err);
-                      }
-                      else{
-                          var request = JSON.parse(res.body);
-                          expect(res.headers['content-type']).to.equal('application/json');
-                          done();
-                      }
-                  });
-        });
-
-
         it ('An LRS sends a header response with "X-Experience-API-Version" as the name and "1.0.3" as the value (Format, 6.2.a, 6.2.b)', function (done){
           this.timeout(0);
           var id = helper.generateUUID();
@@ -3624,20 +3250,6 @@
                     .expect(200, done);
         });
 
-        it('An LRS rejects with error code 400 Bad Request, a Request whose "authority" is a Group and consists of non-O-Auth Agents (4.1.9.a)', function (done) {
-            var templates = [
-                {statement: '{{statements.default}}'},
-                {authority: {"objectType": "Group", "name": "xAPI Group", "mbox": "mailto:xapigroup@example.com",
-                "member":[{"name":"agentA","mbox":"mailto:agentA@example.com"},{"name":"agentB","mbox":"mailto:agentB@example.com"}]}}
-            ];
-            var data = createFromTemplate(templates);
-            data = data.statement;
-            request(helper.getEndpointAndAuth())
-                .post(helper.getEndpointStatements())
-                .headers(helper.addAllHeaders({}))
-                .json(data)
-                .expect(400, done)
-        });
         if(!global.OAUTH)
         {
             //This test appears to only make sense in the case of http basic Auth. Should we have additional tests for bad OAUTH, which is more complicated?
